imporved text report

+++ DuplicateCheck/src/com.dupcheck.strategy/com/dupcheck/strategy/ReportFileStrategy.java
@@ -6,45 +6,55 @@ import java.nio.file.*;
 import java.nio.file.attribute.*;
 public class ReportFileStrategy implements FeatureStrategy {
 	
-	BufferedWriter writer = null;
-	
-	PosixFileAttributes attr = null;
+	private BufferedWriter writer = null;	
 	
 	@Override public void execute(List<File> list){
-		
-		try{
-		writer = new BufferedWriter( new FileWriter( new File("DuplicateFiles.txt"),true));
-		
-		writer.write("FileName\t Owner\tGroup\tSize (bytes)\n");
-		
-		list.forEach(
-			e -> {
-				try{
-				
-				attr = Files.readAttributes(e.toPath() ,PosixFileAttributes.class);
-				
-				writer.write(e.getName()+"\t "+attr.owner().getName()+"\t "+attr.group().getName()+"\t "+e.length());
-				writer.newLine();
-				writer.flush(); 	
-				}catch(IOException ioex){
-					System.out.println("Error occured while writing file.");
-					//ioex.printStackTrace();
-				}		
-			}
-		);
-
-		writer.write("\n\n");
-		
-		}catch(IOException exception){
-			//excecption.printStackTrace();
-			System.out.println("Error occured in creating report file.");
-		}finally{
+	
+		if(list != null ){
+	
 			try{
-			writer.close();
-			}catch(IOException ioex){
-				System.out.println("Error occured in closing object.");
+		
+				if(Files.exists(Paths.get("DuplicateFiles.txt"))){
+					writer = new BufferedWriter( new FileWriter( new File("DuplicateFiles.txt"),true));
+					createReport(writer,list);
+				}else{
+						writer = new BufferedWriter( new FileWriter( new File("DuplicateFiles.txt")));
+						writer.write("FileName\tSize (bytes)\n");
+						writer.write("\n");	
+						writer.flush();
+						createReport(writer,list);
+				}
+			}catch(IOException exception){
+					//excecption.printStackTrace();
+					System.out.println("Error occured in creating file.");
+			}finally{
+					try{
+						writer.close();
+					}catch(IOException ioex){
+						System.out.println("Error occured in closing object.");
+					}
 			}
+		}else{
+			System.out.println("Null argument is passed.");
 		}
 	}
 	
+	
+	private void createReport(BufferedWriter bufferedWriter, List<File> list) throws IOException{
+			list.forEach(
+					e -> {
+							try{			
+								bufferedWriter.write(e.getName()+"\t "+e.length());
+								bufferedWriter.newLine();
+								bufferedWriter.flush(); 	
+							}catch(IOException ioex){
+								System.out.println("Error occured while writing file.");
+								//ioex.printStackTrace();
+							}		
+						}
+					);
+					
+			writer.write("\n");
+					
+		}	
 }
